Use the same bbox object across all cv utils
